trading_ig/requirements.txt
black
certifi
charset-normalizer
click
colorama
idna
munch
mypy-extensions
numpy
pandas==1.4.4
pathspec
patsy
platformdirs
pycryptodome
python-dateutil
python-dotenv
pytz
requests
requests-cache
scipy
six
statsmodels
stocktrends
tenacity
tomli
trading-ig==0.0.15
typing-extensions
urllib3
